Revert "Refactor General Reboot"
This reverts commit the refactor changes, but reworks the face overwrite feature so it can still be used.
